Move podsecuritypolicy registry to policy package.
