+++ pkg/registry/BUILD
@@ -70,11 +70,11 @@ filegroup(
         "//pkg/registry/events/rest:all-srcs",
         "//pkg/registry/extensions/controller/storage:all-srcs",
         "//pkg/registry/extensions/ingress:all-srcs",
-        "//pkg/registry/extensions/podsecuritypolicy:all-srcs",
         "//pkg/registry/extensions/rest:all-srcs",
         "//pkg/registry/networking/networkpolicy:all-srcs",
         "//pkg/registry/networking/rest:all-srcs",
         "//pkg/registry/policy/poddisruptionbudget:all-srcs",
+        "//pkg/registry/policy/podsecuritypolicy:all-srcs",
         "//pkg/registry/policy/rest:all-srcs",
         "//pkg/registry/rbac:all-srcs",
         "//pkg/registry/registrytest:all-srcs",
