+++ pkg/registry/extensions/rest/BUILD
@@ -17,8 +17,8 @@ go_library(
         "//pkg/registry/apps/replicaset/storage:go_default_library",
         "//pkg/registry/extensions/controller/storage:go_default_library",
         "//pkg/registry/extensions/ingress/storage:go_default_library",
-        "//pkg/registry/extensions/podsecuritypolicy/storage:go_default_library",
         "//pkg/registry/networking/networkpolicy/storage:go_default_library",
+        "//pkg/registry/policy/podsecuritypolicy/storage:go_default_library",
         "//vendor/k8s.io/api/extensions/v1beta1:go_default_library",
         "//vendor/k8s.io/apiserver/pkg/registry/generic:go_default_library",
         "//vendor/k8s.io/apiserver/pkg/registry/rest:go_default_library",
