+++ pkg/registry/policy/podsecuritypolicy/doc.go
@@ -16,4 +16,4 @@ limitations under the License.
 
 // Package podsecuritypolicy provides Registry interface and its REST
 // implementation for storing PodSecurityPolicy api objects.
-package podsecuritypolicy // import "k8s.io/kubernetes/pkg/registry/extensions/podsecuritypolicy"
+package podsecuritypolicy // import "k8s.io/kubernetes/pkg/registry/policy/podsecuritypolicy"
