+++ pkg/registry/policy/podsecuritypolicy/storage/BUILD
@@ -27,14 +27,13 @@ go_test(
 go_library(
     name = "go_default_library",
     srcs = ["storage.go"],
-    importpath = "k8s.io/kubernetes/pkg/registry/extensions/podsecuritypolicy/storage",
+    importpath = "k8s.io/kubernetes/pkg/registry/policy/podsecuritypolicy/storage",
     deps = [
         "//pkg/apis/policy:go_default_library",
         "//pkg/printers:go_default_library",
         "//pkg/printers/internalversion:go_default_library",
         "//pkg/printers/storage:go_default_library",
-        "//pkg/registry/extensions/podsecuritypolicy:go_default_library",
-        "//vendor/k8s.io/api/extensions/v1beta1:go_default_library",
+        "//pkg/registry/policy/podsecuritypolicy:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/runtime:go_default_library",
         "//vendor/k8s.io/apiserver/pkg/registry/generic:go_default_library",
         "//vendor/k8s.io/apiserver/pkg/registry/generic/registry:go_default_library",
