+++ pkg/registry/policy/rest/BUILD
@@ -12,8 +12,8 @@ go_library(
     deps = [
         "//pkg/api/legacyscheme:go_default_library",
         "//pkg/apis/policy:go_default_library",
-        "//pkg/registry/extensions/podsecuritypolicy/storage:go_default_library",
         "//pkg/registry/policy/poddisruptionbudget/storage:go_default_library",
+        "//pkg/registry/policy/podsecuritypolicy/storage:go_default_library",
         "//vendor/k8s.io/api/policy/v1beta1:go_default_library",
         "//vendor/k8s.io/apiserver/pkg/registry/generic:go_default_library",
         "//vendor/k8s.io/apiserver/pkg/registry/rest:go_default_library",
