+++ test/test_owners.csv
@@ -749,12 +749,12 @@ k8s.io/kubernetes/pkg/registry/extensions/deployment,dchen1107,1,
 k8s.io/kubernetes/pkg/registry/extensions/deployment/storage,timothysc,1,
 k8s.io/kubernetes/pkg/registry/extensions/ingress,apelisse,1,
 k8s.io/kubernetes/pkg/registry/extensions/ingress/storage,luxas,1,
-k8s.io/kubernetes/pkg/registry/extensions/podsecuritypolicy/storage,dchen1107,1,
 k8s.io/kubernetes/pkg/registry/extensions/replicaset,rrati,0,
 k8s.io/kubernetes/pkg/registry/extensions/replicaset/storage,wojtek-t,1,
 k8s.io/kubernetes/pkg/registry/extensions/rest,rrati,0,
 k8s.io/kubernetes/pkg/registry/policy/poddisruptionbudget,Q-Lee,1,
 k8s.io/kubernetes/pkg/registry/policy/poddisruptionbudget/storage,dchen1107,1,
+k8s.io/kubernetes/pkg/registry/policy/podsecuritypolicy/storage,dchen1107,1,
 k8s.io/kubernetes/pkg/registry/rbac/reconciliation,roberthbailey,1,
 k8s.io/kubernetes/pkg/registry/rbac/validation,rkouj,0,
 k8s.io/kubernetes/pkg/registry/storage/storageclass,brendandburns,1,
